refactor: move public keys to new subpackage
