+++ keys.go
@@ -11,12 +11,12 @@ import (
 	"io"
 	"io/ioutil"
 	"os"
-	"strconv"
 	"time"
 
 	"github.com/go-errors/errors"
 	"github.com/privacybydesign/gabi/big"
 	"github.com/privacybydesign/gabi/internal/common"
+	"github.com/privacybydesign/gabi/keys"
 	"github.com/privacybydesign/gabi/revocation"
 	"github.com/privacybydesign/gabi/safeprime"
 	"github.com/privacybydesign/gabi/signed"
@@ -31,29 +31,23 @@ const (
 
 // PrivateKey represents an issuer's private key.
 type PrivateKey struct {
-	XMLName    xml.Name `xml:"http://www.zurich.ibm.com/security/idemix IssuerPrivateKey"`
-	Counter    uint     `xml:"Counter"`
-	ExpiryDate int64    `xml:"ExpiryDate"`
-	P          *big.Int `xml:"Elements>p"`
-	Q          *big.Int `xml:"Elements>q"`
-	PPrime     *big.Int `xml:"Elements>pPrime"`
-	QPrime     *big.Int `xml:"Elements>qPrime"`
-	order      *big.Int
-	ECDSA      string `xml:",omitempty"`
-
+	keys.PrivateKey
+	order         *big.Int
 	revocationKey *revocation.PrivateKey
 }
 
 // NewPrivateKey creates a new issuer private key using the provided parameters.
 func NewPrivateKey(p, q *big.Int, ecdsa string, counter uint, expiryDate time.Time) *PrivateKey {
 	sk := PrivateKey{
-		P:          p,
-		Q:          q,
-		PPrime:     new(big.Int).Rsh(p, 1),
-		QPrime:     new(big.Int).Rsh(q, 1),
-		Counter:    counter,
-		ExpiryDate: expiryDate.Unix(),
-		ECDSA:      ecdsa,
+		PrivateKey: keys.PrivateKey{
+			P:          p,
+			Q:          q,
+			PPrime:     new(big.Int).Rsh(p, 1),
+			QPrime:     new(big.Int).Rsh(q, 1),
+			Counter:    counter,
+			ExpiryDate: expiryDate.Unix(),
+			ECDSA:      ecdsa,
+		},
 	}
 
 	sk.CacheOrder()
@@ -221,123 +215,28 @@ func GenerateRevocationKeypair(privk *PrivateKey, pubk *PublicKey) error {
 	return nil
 }
 
-// xmlBases is an auxiliary struct to encode/decode the odd way bases are
-// represented in the xml representation of public keys
-type xmlBases struct {
-	Num   int        `xml:"num,attr"`
-	Bases []*xmlBase `xml:",any"`
-}
-
-type xmlBase struct {
-	XMLName xml.Name
-	Bigint  string `xml:",innerxml"` // Has to be a string for ",innerxml" to work
-}
-
-// xmlFeatures is an auxiliary struct to make the XML encoding/decoding a bit
-// easier while keeping the struct for PublicKey somewhat simple.
-type xmlFeatures struct {
-	Epoch struct {
-		Length int `xml:"length,attr"`
-	}
-}
-
-// Bases is a type that is introduced to simplify the encoding/decoding of
-// a PublicKey whilst using the xml support of Go's standard library.
-type Bases []*big.Int
-
-// UnmarshalXML is an internal function to simplify decoding a PublicKey from
-// XML.
-func (bl *Bases) UnmarshalXML(d *xml.Decoder, start xml.StartElement) error {
-	var t xmlBases
-
-	if err := d.DecodeElement(&t, &start); err != nil {
-		return err
-	}
-
-	arr := make([]*big.Int, t.Num)
-	for i := range arr {
-		arr[i], _ = new(big.Int).SetString(t.Bases[i].Bigint, 10)
-	}
-
-	*bl = Bases(arr)
-	return nil
-}
-
-// MarshalXML is an internal function to simplify encoding a PublicKey to XML.
-func (bl *Bases) MarshalXML(e *xml.Encoder, start xml.StartElement) error {
-	l := len(*bl)
-	bases := make([]*xmlBase, l)
-
-	for i := range bases {
-		bases[i] = &xmlBase{
-			XMLName: xml.Name{Local: "Base_" + strconv.Itoa(i)},
-			Bigint:  (*bl)[i].String(),
-		}
-	}
-
-	t := xmlBases{
-		Num:   l,
-		Bases: bases,
-	}
-	return e.EncodeElement(t, start)
-}
-
-// EpochLength is a type that is introduced to simplify the encoding/decoding of
-// a PublicKey whilst using the xml support of Go's standard library.
-type EpochLength int
-
-// UnmarshalXML is an internal function to simplify decoding a PublicKey from
-// XML.
-func (el *EpochLength) UnmarshalXML(d *xml.Decoder, start xml.StartElement) error {
-	var t xmlFeatures
-
-	if err := d.DecodeElement(&t, &start); err != nil {
-		return err
-	}
-	*el = EpochLength(t.Epoch.Length)
-	return nil
-}
-
-// MarshalXML is an internal function to simplify encoding a PublicKey to XML.
-func (el *EpochLength) MarshalXML(e *xml.Encoder, start xml.StartElement) error {
-	var t xmlFeatures
-	t.Epoch.Length = int(*el)
-	return e.EncodeElement(t, start)
-}
-
 // PublicKey represents an issuer's public key.
 type PublicKey struct {
-	XMLName     xml.Name          `xml:"http://www.zurich.ibm.com/security/idemix IssuerPublicKey"`
-	Counter     uint              `xml:"Counter"`
-	ExpiryDate  int64             `xml:"ExpiryDate"`
-	N           *big.Int          `xml:"Elements>n"` // Modulus n
-	Z           *big.Int          `xml:"Elements>Z"` // Generator Z
-	S           *big.Int          `xml:"Elements>S"` // Generator S
-	G           *big.Int          `xml:"Elements>G"` // Generator G for revocation
-	H           *big.Int          `xml:"Elements>H"` // Generator H for revocation
-	R           Bases             `xml:"Elements>Bases"`
-	EpochLength EpochLength       `xml:"Features"`
-	Params      *SystemParameters `xml:"-"`
-	Issuer      string            `xml:"-"`
-	ECDSA       string            `xml:",omitempty"`
-
+	keys.PublicKey
 	revocationKey *revocation.PublicKey
 }
 
 // NewPublicKey creates and returns a new public key based on the provided parameters.
 func NewPublicKey(N, Z, S, G, H *big.Int, R []*big.Int, ecdsa string, counter uint, expiryDate time.Time) *PublicKey {
 	return &PublicKey{
-		Counter:     counter,
-		ExpiryDate:  expiryDate.Unix(),
-		N:           N,
-		Z:           Z,
-		S:           S,
-		R:           R,
-		G:           G,
-		H:           H,
-		EpochLength: DefaultEpochLength,
-		Params:      DefaultSystemParameters[N.BitLen()],
-		ECDSA:       ecdsa,
+		PublicKey: keys.PublicKey{
+			Counter:     counter,
+			ExpiryDate:  expiryDate.Unix(),
+			N:           N,
+			Z:           Z,
+			S:           S,
+			R:           R,
+			G:           G,
+			H:           H,
+			EpochLength: DefaultEpochLength,
+			Params:      DefaultSystemParameters[N.BitLen()],
+			ECDSA:       ecdsa,
+		},
 	}
 }
 
@@ -459,7 +358,7 @@ func (pubk *PublicKey) WriteToFile(filename string, forceOverwrite bool) (int64,
 
 // findMatch returns the first element of safeprimes that makes a suitable pair with p:
 // p*q has the required bith length and p != q mod 8.
-func findMatch(safeprimes []*big.Int, param *SystemParameters, p *big.Int,
+func findMatch(safeprimes []*big.Int, param *keys.SystemParameters, p *big.Int,
 	n, pMod8, qMod8 *big.Int, // temp vars allocated by caller
 ) *big.Int {
 	for _, q := range safeprimes {
@@ -470,7 +369,7 @@ func findMatch(safeprimes []*big.Int, param *SystemParameters, p *big.Int,
 	return nil
 }
 
-func generateSafePrimePair(param *SystemParameters) (*big.Int, *big.Int, error) {
+func generateSafePrimePair(param *keys.SystemParameters) (*big.Int, *big.Int, error) {
 	primeSize := param.Ln / 2
 
 	// Declare and allocate all vars outside the loop and outside the helper function above
@@ -511,24 +410,30 @@ loop: // we need this label to continue the for loop from within the select belo
 }
 
 // GenerateKeyPair generates a private/public keypair for an Issuer
-func GenerateKeyPair(param *SystemParameters, numAttributes int, counter uint, expiryDate time.Time) (*PrivateKey, *PublicKey, error) {
+func GenerateKeyPair(param *keys.SystemParameters, numAttributes int, counter uint, expiryDate time.Time) (*PrivateKey, *PublicKey, error) {
 	p, q, err := generateSafePrimePair(param)
 	if err != nil {
 		return nil, nil, err
 	}
 
 	priv := &PrivateKey{
-		P:          p,
-		Q:          q,
-		PPrime:     new(big.Int).Rsh(p, 1),
-		QPrime:     new(big.Int).Rsh(q, 1),
-		Counter:    counter,
-		ExpiryDate: expiryDate.Unix(),
+		PrivateKey: keys.PrivateKey{
+			P:          p,
+			Q:          q,
+			PPrime:     new(big.Int).Rsh(p, 1),
+			QPrime:     new(big.Int).Rsh(q, 1),
+			Counter:    counter,
+			ExpiryDate: expiryDate.Unix(),
+		},
 	}
 	priv.order = new(big.Int).Mul(priv.PPrime, priv.QPrime)
 
 	// compute n
-	pubk := &PublicKey{Params: param, EpochLength: DefaultEpochLength, Counter: counter, ExpiryDate: expiryDate.Unix()}
+	pubk := &PublicKey{
+		PublicKey: keys.PublicKey{
+			Params: param, EpochLength: DefaultEpochLength, Counter: counter, ExpiryDate: expiryDate.Unix(),
+		},
+	}
 	pubk.N = new(big.Int).Mul(priv.P, priv.Q)
 
 	// Find an acceptable value for S; we follow lead of the Silvia code here:
